+++ android/src/main/java/org/maplibre/reactnative/components/sources/shapesource/ClusterPropertyEntry.kt
@@ -1,0 +1,5 @@
+package org.maplibre.reactnative.components.sources.shapesource
+
+import org.maplibre.android.style.expressions.Expression
+
+class ClusterPropertyEntry(var operator: Expression?, var mapping: Expression?)
