--- android/src/main/java/org/maplibre/reactnative/utils/ClusterPropertyEntry.kt
@@ -1,5 +1,0 @@
-package org.maplibre.reactnative.utils
-
-import org.maplibre.android.style.expressions.Expression
-
-class ClusterPropertyEntry(var operator: Expression?, var mapping: Expression?)
